% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comparison_id.R
\name{add_comparison_id}
\alias{add_comparison_id}
\title{Add Comparison ID}
\usage{
add_comparison_id(data)
}
\arguments{
\item{data}{A data frame containing data for a meta-analysis}
}
\description{
Add a comparison ID (\code{comparison_id}) to keep track of studies in case of
identical \code{study_label}
}
\examples{
iipt_dataset \%>\%
  add_comparison_id()
}
