% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_save.R
\name{batch_save}
\alias{batch_save}
\title{Save Plots in Batch}
\usage{
batch_save(
  data,
  plot_list = plots,
  file_name_list = file_names,
  path,
  width = NA,
  height = NA
)
}
\arguments{
\item{data}{A dataframe containing a list column of plots}

\item{plot_list}{A string indicating the plot column}

\item{file_name_list}{A string indicating the file name column}

\item{path}{A string to a directory where plots should be saved}

\item{width}{Numeric, plot width}

\item{height}{Numeric, plot height}
}
\description{
Save all plots in a list at once to a given directory.
}
