% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baujat_plot.R
\name{baujat_plot}
\alias{baujat_plot}
\title{Baujat Plots}
\source{
\url{https://github.com/cran/metafor/blob/master/R/baujat.rma.r}
}
\usage{
baujat_plot(
  x,
  xlim,
  ylim,
  xlab,
  ylab,
  cex,
  symbol,
  grid = TRUE,
  progbar = FALSE,
  plot_title,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"rma"}.}

\item{xlim}{x-axis limits. If unspecified, the function tries to set the x-axis limits to some sensible values.}

\item{ylim}{y-axis limits. If unspecified, the function tries to set the y-axis limits to some sensible values.}

\item{xlab}{title for the x-axis. If unspecified, the function tries to set an appropriate axis title.}

\item{ylab}{title for the y-axis. If unspecified, the function tries to set an appropriate axis title.}

\item{cex}{optional character expansion factor. If unspecified, the function tries to set this to a sensible value.}

\item{symbol}{either an integer to specify the \code{pch} value (i.e., plotting symbol), or \code{"slab"} to plot the study labels (if specified), or \code{"ids"} to plot the study id numbers (if unspecified, this is the default).}

\item{grid}{logical indicating whether a grid should be added to the plot (can also be a color name).}

\item{progbar}{logical indicating whether a progress bar should be shown (the default is \code{FALSE}).}

\item{plot_title}{A string, the plot title}

\item{...}{other arguments.}
}
\description{
Baujat plots to identify potential study outliers.
}
