% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fixed_meta_analysis.R
\name{calc_fixed_meta_analysis}
\alias{calc_fixed_meta_analysis}
\title{Fixed Effects Meta-Analysis}
\usage{
calc_fixed_meta_analysis(data)
}
\arguments{
\item{data}{A dataframe containing the effect sizes and their associated variance}
}
\description{
Calculate a fixed effects meta-analysis. This is mainly used for Baujat plots
}
