% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_meta_analysis.R
\name{calc_meta_analysis}
\alias{calc_meta_analysis}
\title{Calculate a Three-Level Meta-Analysis}
\usage{
calc_meta_analysis(data, ...)
}
\arguments{
\item{data}{A data frame containing data for a meta-analysis}

\item{...}{Additional options passed to \code{\link[metafor:rma.mv]{metafor::rma.mv()}}}
}
\value{
An object of class \code{rma.mv}
}
\description{
Calculate the three-level meta-analysis with restricted maximum-likelihood and
Knapp-Hartung adjustement for confidence intervals. In some cases, there is
only one \code{study_id} or \code{intervention_id}. In this case, the function would throw a
message, which is suppressed as we know what is happening there.
}
