% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_k_not_one.R
\name{check_k_not_one}
\alias{check_k_not_one}
\title{Check If k Is Not Equal To 1 In All Cases}
\usage{
check_k_not_one(data)
}
\arguments{
\item{data}{A dataframe with column \code{k}}
}
\value{
Logical
}
\description{
This is a helper function needed to plot trajectories of outcomes. If for an
outcome (including all measurements) the number of studies (k) per
measurement is only 1, it should not be used to plot the results because this
plots should only include meta-analytic summary effects.
}
