% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_i_square.R
\name{ext_i_square}
\alias{ext_i_square}
\title{Calculate I Squared}
\source{
\url{https://github.com/MathiasHarrer/dmetar/blob/master/R/mlm.variance.distribution.R}
}
\usage{
ext_i_square(rma_object)
}
\arguments{
\item{rma_object}{Object of class \code{rma.mv} modeling a three-level meta-analysis}
}
\value{
A \link[tibble:tibble-package]{tibble} with 1 row and 4 columns
}
\description{
Calculate I squared for each level of a three-level meta-analysis
}
