% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_meta_data.R
\name{ext_meta_data}
\alias{ext_meta_data}
\title{Extract Meta-Analysis Estimates}
\usage{
ext_meta_data(rma_object)
}
\arguments{
\item{rma_object}{An object of class \code{rma.mv} modelling a three-level meta-analysis}
}
\description{
Extract meta-analysis estimates (I squared, CI, and PI) from an \code{rma.mv} object.
}
