% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_single_data.R
\name{ext_single_data}
\alias{ext_single_data}
\title{Extract estimates for single study comparisons}
\usage{
ext_single_data(data)
}
\arguments{
\item{data}{A dataframe containing the effect size with its associated SE}
}
\description{
Extract estimates for single study comparisons
}
