% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest Plots}
\usage{
forest_plot(data, rma_object, plot_title = NA)
}
\arguments{
\item{data}{A dataframe containing the data}

\item{rma_object}{An object of class \code{rma.mv}}

\item{plot_title}{A string containing the plot title}
}
\description{
Forst plots for given data and rma object.
}
