% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_plot.R
\name{funnel_plot}
\alias{funnel_plot}
\title{Funnel Plot}
\usage{
funnel_plot(
  data,
  plot_title = NA,
  refline = 0,
  alpha = 0.4,
  percent_expansion = 0.05,
  ci_resolution = 1000
)
}
\arguments{
\item{data}{A dataframe containing the effect sizes}

\item{plot_title}{A string for the plot title}

\item{refline}{Numeric, position of reference line (defaults to 0)}

\item{alpha}{Numeric, transparency value (between 0 an 1, defaults to 0.4)}

\item{percent_expansion}{Numeric, value for overplotting to avoid clipping}

\item{ci_resolution}{Numeric, number of points to plot confidence interval}
}
\description{
Funnel plots based on a dataframe.
}
