% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis_without.R
\name{meta_analysis_without}
\alias{meta_analysis_without}
\title{Meta-Analysis without a Particular Study}
\usage{
meta_analysis_without(data, analysis_number, comparison, without, ...)
}
\arguments{
\item{data}{A dataframe containing the effect sizes, CI and analysis numbers}

\item{analysis_number}{Numeric, the analysis number}

\item{comparison}{A string, the measurement against "pre" should be compared}

\item{without}{A string, the study's first author of who the study should be omitted}

\item{...}{Additional options}
}
\description{
Calculate a three-level meta-analysis for a given comparison without a
particular study.
}
