% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot Mean Effects with Confidence Intervals}
\usage{
plot_trajectory(data, analysis_title)
}
\arguments{
\item{data}{A dataframe containing an effect size with associated CI}

\item{analysis_title}{A string used for the plot title}
}
\description{
Plot mean (pooled and single) effects with confidence intervals over time.
}
