% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_logan.R
\name{plot_trajectory_logan}
\alias{plot_trajectory_logan}
\title{Plot Trajectory for Logan et al. Study}
\usage{
plot_trajectory_logan(data, analysis_title)
}
\arguments{
\item{data}{A dataframe containing the effect size and its associated CI}

\item{analysis_title}{A string for the plot title}
}
\description{
In the Logan et al. (2015) Study, the follow-up ranged from 1 - 3 months, so
their results can not be plotted together with all other reults.
}
